Add all optional arguments to listObjectVersions and listObjects
TODO: There is a dependency between versionIdMarker and keyMarker
which cannot be represented correctly with this Devkit version.
A @Required annotation has been created in a snapshot version
which may help solve this problem. This is only to improve UX
in Studio, since there is a descriptive error message at runtime
in case incorrect arguments are provided.

This commit removes the S3TestCase class which often breaks
Maven builds and is not really a part of our test suite.

Also, currently there is no way in Studio to ensure a blank entry
does not appear in the Encoding Type dropdown.